fixed error in module name and refactored errors module
