--- deepparse/errors/data_error/__init__.py
@@ -1,2 +1,0 @@
-# pylint: disable=wildcard-import
-from .data_error import *
